class made, not working
