+++ PRISMRendering/PRISMRenderingPoints/__init__.py
@@ -1,0 +1,1 @@
+from PRISMRenderingPoints.CustomShaderPoints import CustomShaderPoints
